backend/requirements.txt
fastapi[standard]>=0.115.0
python-multipart>=0.0.12
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
pydantic>=2.5.0